Generate .ttf instead of .otf
Building .otf, it shows extra spaces on hankaku glyphs on Windows.  It
is because the font files has a strange value of the xAvgCharWidth
property: 1984.  This commit changes the type of fonts: .otf -> .ttf.
Then fontforge can calculate the valid value for it (1984 -> 1024) and
it shows valid spaces on glyphs.